+++ packages/apollo-language-server/src/project/service.ts
@@ -5,11 +5,6 @@ import { ServiceConfig } from "../config";
 import { ClientIdentity } from "../engine";
 import URI from "vscode-uri";
 
-import {
-  ApolloFederationInfoProvider,
-  federationInfoProviderFromConfig
-} from "../providers/federation-info";
-
 export function isServiceProject(
   project: GraphQLProject
 ): project is GraphQLServiceProject {
@@ -23,8 +18,6 @@ export interface GraphQLServiceProjectConfig {
   loadingHandler: LoadingHandler;
 }
 export class GraphQLServiceProject extends GraphQLProject {
-  public federationInfoProvider: ApolloFederationInfoProvider;
-
   constructor({
     clientIdentity,
     config,
@@ -40,7 +33,6 @@ export class GraphQLServiceProject extends GraphQLProject {
 
     super({ config, fileSet, loadingHandler, clientIdentity });
     this.config = config;
-    this.federationInfoProvider = federationInfoProviderFromConfig(config);
   }
 
   get displayName() {
@@ -58,6 +50,6 @@ export class GraphQLServiceProject extends GraphQLProject {
   }
 
   resolveFederationInfo() {
-    return this.federationInfoProvider.resolveFederationInfo();
+    return this.schemaProvider.resolveFederatedServiceSDL();
   }
 }