+++ packages/apollo/src/commands/service/__tests__/check.test.ts
@@ -134,63 +134,6 @@ function mockIntrospectionQuery() {
 }
 
 /**
- * Mock the network requests for a composition failure.
- */
-function mockCompositionFailure() {
-  mockIntrospectionQuery();
-
-  nock(localURL, {
-    encodedQueryParams: true
-  })
-    .post(
-      "/graphql",
-      ({ operationName }) => operationName === "getFederationInfo"
-    )
-    .reply(200, {
-      data: {
-        _service: {
-          sdl:
-            'extend type Query {\n  me: User\n}\n\ntype User @key(fields: "id") {\n  name: String\n  username: String\n  birthDate: String\n}\n'
-        }
-      }
-    });
-
-  nock("https://engine-staging-graphql.apollographql.com:443", {
-    encodedQueryParams: true
-  })
-    .post(
-      "/api/graphql",
-      ({ operationName }) => operationName === "CheckPartialSchema"
-    )
-    .reply(200, {
-      data: {
-        service: {
-          validatePartialSchemaOfImplementingServiceAgainstGraph: {
-            compositionValidationDetails: {
-              schemaHash: null
-            },
-            warnings: [],
-            errors: [
-              {
-                message:
-                  "[reviews] User.id -> marked @external but it does not have a matching field on on the base service (accounts)"
-              },
-              {
-                message:
-                  "[reviews] User -> A @key selects id, but User.id could not be found"
-              },
-              {
-                message:
-                  "[accounts] User -> A @key selects id, but User.id could not be found"
-              }
-            ]
-          }
-        }
-      }
-    });
-}
-
-/**
  * Mock network requests for a successful schema composition. This includes the subsequent `CheckSchema`
  * request that will be made.
  */
@@ -223,13 +166,39 @@ function mockCompositionSuccess() {
     .reply(200, {
       data: {
         service: {
-          validatePartialSchemaOfImplementingServiceAgainstGraph: {
-            compositionValidationDetails: {
-              schemaHash:
-                "645fdd4b789fffb5c5b59443a12e6f575e61345e95fe9e1dae3fe9acb23c68efa8ac31ea657892f0a85d1c90d8503fe9e482f520fe8d9786ae26948de10ce4a6"
+          checkPartialSchema: {
+            compositionValidationResult: {
+              compositionValidationDetails: {
+                schemaHash:
+                  "645fdd4b789fffb5c5b59443a12e6f575e61345e95fe9e1dae3fe9acb23c68efa8ac31ea657892f0a85d1c90d8503fe9e482f520fe8d9786ae26948de10ce4a6"
+              },
+              graphCompositionID: null,
+              errors: []
             },
-            warnings: [],
-            errors: []
+            checkSchemaResult: {
+              targetUrl:
+                "https://engine-staging.apollographql.com/service/justin-fullstack-tutorial/check/3acd7765-61b2-4f1a-9227-8b288e42bfdc",
+              diffToPrevious: {
+                severity: "NOTICE",
+                affectedClients: [],
+                affectedQueries: [],
+                numberOfCheckedOperations: 0,
+                changes: [
+                  {
+                    severity: "NOTICE",
+                    code: "ARG_CHANGED_TYPE",
+                    description:
+                      "`Query.launches` argument `after` has changed type from `String` to `String!`"
+                  }
+                ],
+                validationConfig: {
+                  from: "-47347200",
+                  to: "-0",
+                  queryCountThreshold: 1,
+                  queryCountThresholdPercentage: 0
+                }
+              }
+            }
           }
         }
       }
@@ -358,6 +327,66 @@ function mockNonFederatedSuccess() {
     });
 }
 
+/**
+ * Mock network requests for a federated schema running partialSchemaCheck and producing errors
+ */
+const mockPartialSchemaCheckFailure = () => {
+  mockIntrospectionQuery();
+
+  nock(localURL, {
+    encodedQueryParams: true
+  })
+    .post(
+      "/graphql",
+      ({ operationName }) => operationName === "getFederationInfo"
+    )
+    .reply(200, {
+      data: {
+        _service: {
+          sdl:
+            'extend type Query {\n  me: User\n}\n\ntype User @key(fields: "id") {\n  name: String\n  username: String\n  birthDate: String\n}\n'
+        }
+      }
+    });
+
+  nock("https://engine-staging-graphql.apollographql.com:443", {
+    encodedQueryParams: true
+  })
+    .post(
+      "/api/graphql",
+      ({ operationName }) => operationName === "CheckPartialSchema"
+    )
+    .reply(200, {
+      data: {
+        service: {
+          checkPartialSchema: {
+            compositionValidationResult: {
+              compositionValidationDetails: {
+                schemaHash: null
+              },
+              graphCompositionID: null,
+              errors: [
+                {
+                  message:
+                    "[reviews] User.id -> marked @external but it does not have a matching field on on the base service (accounts)"
+                },
+                {
+                  message:
+                    "[reviews] User -> A @key selects id, but User.id could not be found"
+                },
+                {
+                  message:
+                    "[accounts] User -> A @key selects id, but User.id could not be found"
+                }
+              ]
+            },
+            checkSchemaResult
+          }
+        }
+      }
+    });
+};
+
 describe("service:check", () => {
   let originalChalkEnabled;
 
@@ -404,7 +433,7 @@ describe("service:check", () => {
       describe("should report composition errors correctly", () => {
         it("vanilla", async () => {
           captureApplicationOutput();
-          mockCompositionFailure();
+          mockPartialSchemaCheckFailure();
 
           expect.assertions(2);
 
@@ -422,7 +451,7 @@ describe("service:check", () => {
 
         it("compacts output in CI", async () => {
           captureApplicationOutput();
-          mockCompositionFailure();
+          mockPartialSchemaCheckFailure();
 
           expect.assertions(2);
 
@@ -442,7 +471,7 @@ describe("service:check", () => {
 
         it("--markdown", async () => {
           captureApplicationOutput();
-          mockCompositionFailure();
+          mockPartialSchemaCheckFailure();
 
           expect.assertions(2);
 
@@ -462,7 +491,7 @@ describe("service:check", () => {
 
         it("--json", async () => {
           captureApplicationOutput();
-          mockCompositionFailure();
+          mockPartialSchemaCheckFailure();
 
           expect.assertions(2);